requirements.txt
fastapi
uvicorn
python-multipart
psutil
pytesseract
pymupdf
python-docx
pillow
xlrd
striprtf
odfpy
pandas
openpyxl
tqdm
ollama
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sqlmodel
msgraph-sdk
azure-identity
presidio-analyzer
presidio-anonymizer
spacy
pywin32;platform_system=="Windows"
